minor changes
git-svn-id: https://svn.ufz.de/svn/ogs/trunk/sources@6248 3895a583-e809-0410-9bb6-aa9463026377

+++ FEM/par_ddc.cpp
@@ -49,7 +49,7 @@ using namespace std;
 #include "equation_class.h"
 #endif
 vector<CPARDomain*>dom_vector;
-vector<int> node_connected_doms; //This will be removed after sparse class is finished WW
+vector<long> node_connected_doms; //This will be removed after sparse class is finished WW
 
 
 /**************************************************************************
